Separate VerifierRuleTest and ErrorCollectorTest
It is a common practice to collect the tests for a class in a test class
whose name is equal to the class' name but with a suffix `Test`.
Therefore having all tests for `ErrorCollector` in a class named
`ErrorCollectorTest` makes it easier for people to find them (principle
of least astonishment).
